refactor: replace defaults with dynamic sys-values
In order to make setting the ideal values based on the current system, static defaults are replaced with dynamic calculations for the current system based on its CPU-cores:
- Single process threshold: 25% of total system capacity, clamped to [min: 120%, max: 400%]
- Overall system threshold: 85% of total system capacity
- Check interval: 60s per core, clamp to [min: 1minute, max: 10minute]

These evaluted values are also shown in the setup as a reference point for the user. Further, upon completing the setup, the user-defined values are matched again to the %-share of the user's system. Example:

```
Alerting Thresholds:
- Single Process % CPU usage threshold: 90% (7.5% of total capacity)
- Overall System % CPU threshold: 100% (8.3% of total capacity)
- Check interval: Every 10 seconds (0 minutes)
```